Создаёт модуль task.js
